+++ impact/model/impact_model.py
@@ -4,7 +4,11 @@ from lume.model import LUMEModel
 from lume.variables import Variable
 
 from impact.model.transformer.base import Transformer
-from impact.model.config import VariableMappingConfig, make_transformer, make_variables
+from impact.model.impact_config import (
+    VariableMappingConfig,
+    make_transformer,
+    make_variables,
+)
 
 
 class LUMEImpactModel(LUMEModel):
